Refactor: a less verbose ModalProvider
