+++ src/components/TxFlow/common/TxLayout.tsx
@@ -4,6 +4,7 @@ import TxStatusWidget from '@/components/TxFlow/TxStatusWidget'
 import { type TransactionSummary } from '@safe-global/safe-gateway-typescript-sdk'
 import css from '@/components/TxFlow/TokenTransfer/styles.module.css'
 import { ProgressBar } from '@/components/common/ProgressBar'
+import SafeTxProvider from '../SafeTxProvider'
 import TxNonce from './TxNonce'
 
 const TxLayout = ({
@@ -20,31 +21,33 @@ const TxLayout = ({
   progress?: number
 }) => {
   return (
-    <Container>
-      <Grid container className={css.wrapper} alignItems="center" justifyContent="center">
-        <Grid item xs={12}>
-          <Typography variant="h3" component="div" fontWeight="700" mb={2}>
-            {title}
-          </Typography>
-        </Grid>
-
-        <Grid item container xs={12} gap={3}>
-          <Grid item xs={7} component={Paper}>
-            {progress !== undefined && <ProgressBar value={progress} />}
-
-            <Box display="flex" justifyContent="flex-end" py={2} px={3}>
-              <TxNonce />
-            </Box>
-
-            {children}
+    <SafeTxProvider>
+      <Container>
+        <Grid container className={css.wrapper} alignItems="center" justifyContent="center">
+          <Grid item xs={12}>
+            <Typography variant="h3" component="div" fontWeight="700" mb={2}>
+              {title}
+            </Typography>
           </Grid>
 
-          <Grid item xs={4}>
-            <TxStatusWidget step={step} txSummary={txSummary} />
+          <Grid item container xs={12} gap={3}>
+            <Grid item xs={7} component={Paper}>
+              {progress !== undefined && <ProgressBar value={progress} />}
+
+              <Box display="flex" justifyContent="flex-end" py={2} px={3}>
+                <TxNonce />
+              </Box>
+
+              {children}
+            </Grid>
+
+            <Grid item xs={4}>
+              <TxStatusWidget step={step} txSummary={txSummary} />
+            </Grid>
           </Grid>
         </Grid>
-      </Grid>
-    </Container>
+      </Container>
+    </SafeTxProvider>
   )
 }
