--- src/components/TxFlow/index.ts
@@ -1,7 +1,0 @@
-import RejectTx from '@/components/TxFlow/RejectTx'
-import ReplaceTxMenu from '@/components/TxFlow/ReplaceTx'
-import TokenTransferFlow from '@/components/TxFlow/TokenTransfer/TokenTransferFlow'
-import NewTxMenu from '@/components/TxFlow/NewTx'
-import ConfirmProposedTx from '@/components/TxFlow/ConfirmTx'
-
-export { RejectTx, ReplaceTxMenu, TokenTransferFlow, NewTxMenu, ConfirmProposedTx }
